Separated task details and questions list in task editor
